Refactor crazy fun namespace 1 (#7667)
* Namespaced and fixed up common.rb and export.rb
* Sort out folder and python files
