+++ rake_tasks/crazy_fun/mappings/export_mappings.rb
@@ -1,0 +1,7 @@
+class ExportMappings
+  def add_all(fun)
+    fun.add_mapping("export_file", Export::CheckPreconditions.new)
+    fun.add_mapping("export_file", Export::CreateTask.new)
+    fun.add_mapping("export_file", Export::AddDependencies.new)
+  end
+end
