+++ rake_tasks/crazy_fun/mappings/python/py_task.rb
@@ -1,0 +1,11 @@
+module Python
+  class PyTask < Tasks
+    def get_resources(browser, args)
+      resources = []
+      resources.concat(args[:resources]) if args[:resources]
+      browser_specific_resources = SeleniumRake::Browsers::BROWSERS[browser][:python][:resources]
+      resources.concat(browser_specific_resources) if browser_specific_resources
+      return resources
+    end
+  end
+end
