+++ rake_tasks/crazy_fun/mappings/python_mappings.rb
@@ -1,0 +1,6 @@
+class PythonMappings
+  def add_all(fun)
+    fun.add_mapping("py_test", Python::AddDependencies.new)
+    fun.add_mapping("py_test", Python::RunTests.new)
+  end
+end
